# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/vehicles": {
      "get": {
        "operationId": "Fleet_vehicles",
        "parameters": [
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {},
            "example": "50"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {},
            "example": "1"
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {},
            "example": ""
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Retrieve vehicles",
        "description": "Retrieve operation in the Vehicles module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Fleet_createVehicle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              },
              "example": {
                "name": "سيارة توزيع 01",
                "plateNumber": "ABC 1234",
                "vin": "WVWZZZ1JZXW000001",
                "group": "أسطول الرياض",
                "make": "Toyota",
                "model": "Hilux",
                "year": 2025,
                "color": "#ffffff"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "سيارة توزيع 01",
                    "plateNumber": "ABC 1234",
                    "vin": "WVWZZZ1JZXW000001",
                    "group": "أسطول الرياض",
                    "make": "Toyota",
                    "model": "Hilux",
                    "year": 2025,
                    "color": "#ffffff"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Create / execute vehicles",
        "description": "Create / execute operation in the Vehicles module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/vehicles/{id}": {
      "get": {
        "operationId": "Fleet_vehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Retrieve vehicles",
        "description": "Retrieve operation in the Vehicles module. Organization scope is enforced from the authenticated session."
      },
      "patch": {
        "operationId": "Fleet_updateVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              },
              "example": {
                "name": "سيارة توزيع 01 - محدثة",
                "group": "أسطول جدة",
                "driverId": "driver-id"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "سيارة توزيع 01 - محدثة",
                    "group": "أسطول جدة",
                    "driverId": "driver-id"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Update vehicles",
        "description": "Update operation in the Vehicles module. Organization scope is enforced from the authenticated session."
      },
      "delete": {
        "operationId": "Fleet_deleteVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Delete vehicles",
        "description": "Delete operation in the Vehicles module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/vehicles/{id}/positions": {
      "post": {
        "operationId": "Fleet_position",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDto"
              },
              "example": {
                "latitude": 24.7136,
                "longitude": 46.6753,
                "speed": 73,
                "heading": 120,
                "altitude": 620,
                "ignition": true,
                "accuracy": 5,
                "recordedAt": "2026-08-17T11:00:00.000Z",
                "raw": {
                  "source": "tcp-adapter"
                }
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "latitude": 24.7136,
                    "longitude": 46.6753,
                    "speed": 73,
                    "heading": 120,
                    "altitude": 620,
                    "ignition": true,
                    "accuracy": 5,
                    "recordedAt": "2026-08-17T11:00:00.000Z",
                    "raw": {
                      "source": "tcp-adapter"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Telemetry"
        ],
        "summary": "Create / execute telemetry",
        "description": "Create / execute operation in the Telemetry module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/vehicles/{id}/history": {
      "get": {
        "operationId": "Fleet_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-01T00:00:00.000Z"
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-17T23:59:59.999Z"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Telemetry"
        ],
        "summary": "Retrieve telemetry",
        "description": "Retrieve operation in the Telemetry module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/drivers": {
      "get": {
        "operationId": "Fleet_drivers",
        "parameters": [
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {},
            "example": "50"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {},
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Drivers"
        ],
        "summary": "Retrieve drivers",
        "description": "Retrieve operation in the Drivers module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Fleet_createDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              },
              "example": {
                "name": "خالد محمد",
                "phone": "+966500000000",
                "email": "driver@example.com",
                "licenseNumber": "LIC-1001",
                "licenseExpiresAt": "2028-12-31T00:00:00.000Z",
                "active": true
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "خالد محمد",
                    "phone": "+966500000000",
                    "email": "driver@example.com",
                    "licenseNumber": "LIC-1001",
                    "licenseExpiresAt": "2028-12-31T00:00:00.000Z",
                    "active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Drivers"
        ],
        "summary": "Create / execute drivers",
        "description": "Create / execute operation in the Drivers module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/drivers/{id}": {
      "patch": {
        "operationId": "Fleet_updateDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              },
              "example": {
                "name": "خالد محمد",
                "phone": "+966511111111",
                "licenseNumber": "LIC-1001",
                "active": true
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "خالد محمد",
                    "phone": "+966511111111",
                    "licenseNumber": "LIC-1001",
                    "active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Drivers"
        ],
        "summary": "Update drivers",
        "description": "Update operation in the Drivers module. Organization scope is enforced from the authenticated session."
      },
      "delete": {
        "operationId": "Fleet_deleteDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Drivers"
        ],
        "summary": "Delete drivers",
        "description": "Delete operation in the Drivers module. Organization scope is enforced from the authenticated session."
      }
    },
...
  },
  "info": {
    "title": "ORBIT Fleet API",
    "description": "Complete fleet operations API: authentication, vehicles, TCP-normalized telemetry, drivers, geofences, alerts, maintenance, fuel, tasks, device commands, cameras, configuration, users, audit and reports.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token returned by POST /v1/auth/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly browser session cookie"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "required": [
          "name",
          "plateNumber"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "سيارة توزيع 01"
          },
          "plateNumber": {
            "type": "string",
            "example": "ABC 1234"
          },
          "vin": {
            "type": "string",
            "example": "WVWZZZ1JZXW000001"
          },
          "group": {
            "type": "string",
            "example": "أسطول الرياض"
          },
          "make": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Hilux"
          },
          "year": {
            "type": "integer",
            "minimum": 1980,
            "maximum": 2100,
            "example": 2025
          },
          "color": {
            "type": "string",
            "example": "#ffffff"
          },
          "driverId": {
            "type": "string",
            "example": "driver-id"
          }
        }
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "سيارة توزيع 01 - محدثة"
          },
          "plateNumber": {
            "type": "string",
            "example": "ABC 1234"
          },
          "vin": {
            "type": "string"
          },
          "group": {
            "type": "string",
            "example": "أسطول جدة"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "minimum": 1980,
            "maximum": 2100
          },
          "color": {
            "type": "string"
          },
          "driverId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PositionDto": {
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "speed",
          "heading",
          "ignition",
          "recordedAt"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "minimum": -90,
            "maximum": 90,
            "example": 24.7136
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "minimum": -180,
            "maximum": 180,
            "example": 46.6753
          },
          "speed": {
            "type": "number",
            "minimum": 0,
            "example": 73,
            "description": "Kilometers per hour"
          },
          "heading": {
            "type": "number",
            "minimum": 0,
            "maximum": 360,
            "example": 120
          },
          "altitude": {
            "type": "number",
            "example": 620
          },
          "ignition": {
            "type": "boolean",
            "example": true
          },
          "accuracy": {
            "type": "number",
            "minimum": 0,
            "example": 5
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-17T11:00:00.000Z"
          },
          "raw": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "source": "tcp-adapter"
            }
          }
        }
      },
      "CreateDriverDto": {
        "type": "object",
        "required": [
          "name",
          "licenseNumber"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "خالد محمد"
          },
          "phone": {
            "type": "string",
            "example": "+966500000000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "driver@example.com"
          },
          "licenseNumber": {
            "type": "string",
            "example": "LIC-1001"
          },
          "licenseExpiresAt": {
            "type": "string",
            "format": "date-time",
            "example": "2028-12-31T00:00:00.000Z"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "licenseNumber": {
            "type": "string"
          },
          "licenseExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}